+++ src/hooks/useCheckout.js
@@ -1,0 +1,20 @@
+import { useEffect } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { onAuthStateChanged } from 'firebase/auth'
+import { FirebaseAuth } from '../firebase'
+import { login } from '../store'
+
+export const useCheckout = () => {
+  const { status } = useSelector((state) => state.auth)
+  const dispatch = useDispatch()
+
+  useEffect(() => {
+    onAuthStateChanged(FirebaseAuth, (user) => {
+      if (!user) return dispatch(logout())
+      const { uid, email, displayName, photoURL } = user
+      dispatch(login({ uid, email, displayName, photoURL }))
+    })
+  }, [])
+
+  return status 
+}
