+++ src/router/AppRouter.jsx
@@ -1,24 +1,11 @@
-import { useDispatch, useSelector } from 'react-redux'
-import { onAuthStateChanged } from 'firebase/auth'
-import { useEffect } from 'react'
 import { Navigate, Route, Routes } from 'react-router-dom'
 import { AuthRoutes } from '../auth'
-import { FirebaseAuth } from '../firebase'
+import { useCheckout } from '../hooks'
 import { JournalRoutes } from '../journal/routes/JournalRoutes'
 import { CheckingAuth } from '../ui/components'
-import { login, logout } from '../store'
 
 export const AppRouter = () => {
-  const { status } = useSelector((state) => state.auth)
-  const dispatch = useDispatch()
-
-  useEffect(() => {
-    onAuthStateChanged(FirebaseAuth, (user) => {
-      if (!user) return dispatch(logout())
-      const { uid, email, displayName, photoURL } = user
-      dispatch(login({ uid, email, displayName, photoURL }))
-    })
-  }, [])
+  const status = useCheckout()
 
   if (status === 'checking') {
     return <CheckingAuth />
@@ -31,7 +18,7 @@ export const AppRouter = () => {
         <Route path='/auth/*' element={<AuthRoutes />} />
       )}
 
-      <Route path='/*' element={<Navigate to="/auth/login" />} />
+      <Route path='/*' element={<Navigate to='/auth/login' />} />
       {/* <Route path='/auth/*' element={<AuthRoutes />} />
       <Route path='/*' element={<JournalRoutes />} /> */}
     </Routes>
